This is the 0.07 release of MOOSE Date: 6 September 2006.
Overview: This is actually several releases later than the 0.06 release,
but functionally the major change is the implementation of a solver API.
The basic scheme is 
Numerics <====> Data <====> Hub <====> MOOSE objects.
This can be seen in action in the kinetics directory, where we have
a trivial ForwardEuler solver handling a Stoich object that manages the
data. The communication between the data and the molecules and reactions is
handled by the KineticHub. 
Behind the scenes, the major change is the improvement in the preprocessor.
Most objects can be almost entirely specified using the preprocessor and
again the kinetics directory is the examplar. Have a look at 
DOCS/UsingPreprocessor for its features and remaining limitations.

Compilation instructions:
Edit Makefile to select GCC3 or GCC4 options. Some of the later GCC3 versions
actually seem to prefer the GCC4 flags.
Go into the moose directory and type
        make moose

Currently works on Debian/Ubuntu/Gcc4 and Gcc3. 
Known compilation issues remain with Fedora GCC4.
Linux version 2.6.15-26-686 (buildd@terranova) 
(gcc version 4.0.3 (Ubuntu 4.0.3-1ubuntu5)
-- Upi



git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@6 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: e1318925a0a3c7f832af7717aa459b63d0656e35