With this revision we plan to put all updates onto the SourceForge Subversion tree for MOOSE, rather than waiting for occasional relatively stable releases.
Recent updates for the past week have been the implementation of
LookupFinfo and LookupFtype, which facilitate access to indexed fields
including but not limited to arrays and maps. This was needed
to feed into some new Shell commands for traversing element hierarchy.

All this has been set up to work with SWIG, with a bare minimum
of commands for 'create' and 'le' there to start testing.

To compile for Python and Swig, use the command
	make pymoose
and it will set up the pymoose module for loading into Python.
The regular compile 
	make
will give the usual version of MOOSE which runs a whole lot of
unit tests.

-- Upi



git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@28 d1858a74-f715-0410-99c9-c66fe9cce3d6