WildCard.cpp has been modified to return list of Id for the path given instead of list of Element pointers. Corresponding changes have been made in Shell.cpp. setfield and getfield commands' implementation has been modified to handle arrayelement with indices. A design question..Suppose c is a Comparment array. setfield c x 100. What should this command do? Does it make change in all the indices of c or only in c[0]. In genesis, its the latter, which I feel is useless. This particular thing has not been addressed in this commit. I will address it in next one.
NOTE: Stoich::localSetPath: The WildCard now return a vector of Ids rather than a vector of Elements. So I have permeated the changes to Stoich::rebuildMatrix. This function assumes SimpleElements (or so it seemed like). I have converted the vector <Element *> ret to vector <Id> ret and corresponding converted vector< Element* >::iterator i to vector< Id >::iterator i and *i to ( *i )(). This will not work in all case. One case being, when the id corresponds to a particular index of ArrayElement. I have not dealt with this problem. I think that this problem would repeat and we should stop using Element * entirely and use Eref instead. I hope it is complete the picture. Similar changes have been made in KineticManager::estimateDt
-RD


git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@459 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 550cbc711fdd52f93af98023c669f4c3fcd32dea
Former-commit-id: edd41e9cb9ee2254348c8d252a36722629f445d9
Former-commit-id: cb5658f99710ba250181a0cbb1d2d202ba68c4cb