Numerous fixes in PyMOOSE and moosegui. 1. PyMooseContext::getFieldList was not clearing the vector of field names - resulting in the field list being appended at every call.
2. moosegui - object editor now updates the plot column when a field
is dragged and dropped on the plot window.

3. moosegui - clean up of new object creation. All the list widgets in
MooseClasses widget now escalate the itemDoubleCLicked event via the
MooseClasses via the method signalItemDoubleClicked. This connects to
the ModelTree's insertMooseObjectSlot.

4. moosegui - when a new object is inserted into the tree, the
objecteditor comes up automatically.

5. moosegui - when a field is queued for plotting, the recording-table
name and as a result the legend item is full path of the original
object with '/' replaced by '_' and the field name appended with
another '_' in between.


git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@2176 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 07a39396b75df593f872867b6aaa0c7a649b1f5c
Former-commit-id: 1efc0cfbb82182bc1e6c869dbe0a1b37b2c52608
Former-commit-id: 41221be20cdbab92757debe8f7f0a810761bf06c