Numerous fixes in PyMOOSE and moosegui. 1. PyMooseContext::getFieldList was not clearing the vector of field names - resulting in the field list being appended at every call.
2. moosegui - object editor now updates the plot column when a field
is dragged and dropped on the plot window.

3. moosegui - clean up of new object creation. All the list widgets in
MooseClasses widget now escalate the itemDoubleCLicked event via the
MooseClasses via the method signalItemDoubleClicked. This connects to
the ModelTree's insertMooseObjectSlot.

4. moosegui - when a new object is inserted into the tree, the
objecteditor comes up automatically.

5. moosegui - when a field is queued for plotting, the recording-table
name and as a result the legend item is full path of the original
object with '/' replaced by '_' and the field name appended with
another '_' in between.


git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@2176 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: f74278d79bb47384e6f7ed022e8315ac2c308eb0
Former-commit-id: bc14307e9ff9bcc57fd377713c4bbea1392460b4
Former-commit-id: ef7a055611d5c5197d67824f1783f5d2da5c7adc