The python API slightly changed:
element is a function implemented in C. It can take a path string, an
existing element or an ematrix and return an instance of the
appropriate moose builtin class.

ObjId/element is renamed "melement". This should not be directly
needed by the users. Python code that used element() earlier will
remain valid.

to_el becomes redundant with this change.


git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/buildQ@3926 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 9b40b719531ae0997b75a8e523cd291f8d8aa8e5
Former-commit-id: fc335e9d45c8fc6b300f794f9560853dae43aa54