The python API slightly changed:
element is a function implemented in C. It can take a path string, an
existing element or an ematrix and return an instance of the
appropriate moose builtin class.

ObjId/element is renamed "melement". This should not be directly
needed by the users. Python code that used element() earlier will
remain valid.

to_el becomes redundant with this change.


git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/buildQ@3926 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 8b678e708f4d95701338eeae9c88643d12dd2569