Fixed adjust_chanlib bug in traub_2005.
In summary: my code was designed to update the channel prototypes in
/library before loading .p file for each cell type. The code to adjust
the channel properties was trying to check if the channel is an
instance of KChannel or NaChannel, etc. But the prototypes themselves
were plain HHChannel. Hence they were really not getting updated. Now
I am updating the protype dict with prototype channels wrapped in
correct subclass.


git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/buildQ@4341 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 725e92ea8e8fbcba8d71c8e624ddc8ba7e0c4b41
Former-commit-id: 9d4d4047631a64e44537414d17eaa754f9c6243d