Add license text, clarify that GPLv3+ is the license
The README already said that moose is licensenced under GPLv3, but
didn't specify "or any later version", like headers in the files.
Clarify that the license is GPLv3+, remove copyleft file which
specified LGPL.

Add the text of the license; various Linux distributions expect this
in packages.

The text is taken from http://www.gnu.org/licenses/gpl.txt.