pymoose: simplify and fix arguments to PyArg_ParseTupleAndKeywords
test_pymoose was failing semi-randomly on the line
   oidlist = [moose.Neutral('x_%d' % (ii), n=2**10) for ii in range(2**6)]
which suggests memory corruption.

In moose_ObjId_init_from_id PyArg_ParseTupleAndKeywords was called with
not enough variable specifiers for the provided format string.

Also, the arguments were specified in an overcomplicated way. It's
easier and more efficient to declare the kwlist as a const char* const
array, and to cast it to char**. Python will not modify this array.
With this simplification it is simpler to see if the format string
matches with the keyword list and the number of provided variables.

This change makes the test not fail in the same place, so I assume it
fixes something.