Remove sys.path additions from tests/
When a separate build dir is used, there's no fixed relationship between
the directory that can be added to sys.path and the location of the tests file.
In particular, ../../python will often refer to the source directory, which
does not contain _moose.so and cannot be successfully imported.

for i in ../tests/*/*.py; do PYTHONPATH=python python3 $i; done
- test_moogli.py fails because moogli is not present
- test_sbml.py fails because expected tables are not found, I think this is unrelated
- other tests pass