Updated python setup script to include streaming classes
The Streamer class was introduced in commit 481d46a and the python setup script was failing because Table depended on these and some of the utility functions for this.
Also renamed setup_subha.py to setup.cygwin.py to better reflect its use. This script should be platform independent, but so far it is necessary in cygwin.