Close #169. Renamed MOOSE_VERSION to _MOOSE_VERSION.
Since only developers uses it, it is safe to leave it undocumented in
CMakeLists.txt.

This variable is automatically set by git when building from git
repisitory. When git is not found, then one needs to pass this
information from command line.

This may break the build in `moose-package` repository in future.

Linux builds are safe on Open Build Service since cmake available on
those platforms are now new enough.