Script testing DifShells
Script creates a test model of a dendrite segment with one Ca
channel. DifShells are created. DifBuffers are created. DifShells are
connected to provide diffusion. DifBuffers are connected to provide
diffusion. DifBuffers are connected with DifShells to provide
buffer. MMPumps are created and connected to the outer DifShell. Ca
channel is connected to the outer Difshell to provide Ca influx and
outer DifShell is connected to Ca channel to provide ion
concentration. Segment's Vm, channel's Gk and Ik, DifShell's C and
DifBuffer's bBound are connected with output to save simulation
results.