The default values changed in the commit (ba1065c) breaks example such as https://github.com/BhallaLab/moose-examples/blob/09ca09e68f163a0920653a9400214aff10b6ba16/snippets/rxdFuncDiffusion.py . The sparse matrix has upper limit of number of rows and columns to 20,000. In the example linked, new values makes it impossible to run the example as is.
To overcome this, we notify the user whenever this limit is reached and
ask them to set `diffLength` before setting up the `x1`.

On top of this, we can also revert default values to original larger
values. Not sure why these values were changed to the scale of 1e-6?
Unless the reason is known, I am not going to change current values. The
above linked example  now works with this commit.