Fixed showmsg() bug in pybind11 based version. (#466)
* Fixed showmsg() bug in pybind11 based version.

The newer implementation of showmsg() shows the outgoing messages with
misdirected arrows. Fixed now.

* Added newer version of pybind11

* Swapped clocks for channel and compartment to fix  #467

* This actually fixes scheduling for channels and compartments #467

The previous commit only updated the docs. This one modifies the
actual tick assignments.
Now electrical compartments are scheduled on ticks 2 and 3 and
channels (all derived from ChanBase) on 4.

* Minor fix for compiler warning on loop var creating copy from const

* Removed CMake generated compile command.

the compile command file is generated by cmake in each build
environment and should not be part of the repo.